Fixed bg color changing

+++ 3D_Engine_Core/include/engine/render/open_gl/renderer_open_gl.hpp
@@ -30,7 +30,7 @@ namespace engine::render::open_gl
 		static bool init_with_glfw() noexcept;
 		static void draw(const vertex_array& _vertex_array_buffer, DrawingMode _drawing_mode = DrawingMode::Triangle) noexcept;
 		static void clear(Mask _mask_type) noexcept;
-		static void setClearColor(uint8_t _red, uint8_t _green, uint8_t _blue, float _alpha) noexcept;
+		static void setClearColor(float _red, float _green, float _blue, float _alpha) noexcept;
 		static void setViewport(uint16_t _width, uint16_t _height,
 								uint16_t _left_offset = 0, uint16_t _bottom_offset = 0) noexcept;
